added product component and counter
