--- src/App.js
@@ -1,13 +1,0 @@
-import Products from "./components/Products.jsx";
-import CreateProduct from "./components/CreateProduct.jsx";
-
-function App() {
-  return (
-    <>
-      <Products />
-      <CreateProduct />
-    </>
-  );
-}
-
-export default App;
