+++ src/index.js
@@ -1,15 +1,11 @@
 import React, { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { Provider } from 'react-redux';
-import { store } from './store';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <StrictMode>
-    <Provider store={store} >
-      <App />
-    </Provider>
+    <App />
   </StrictMode>
 );
