--- src/store.js
@@ -1,14 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit'
-import { setupListeners } from '@reduxjs/toolkit/query'
-import { productApi } from './api/products'
-
-export const store = configureStore({
-    reducer: {
-        [productApi.reducerPath]: productApi.reducer,
-    },
-
-    middleware: (getDefaultMiddleware) =>
-        getDefaultMiddleware().concat(productApi.middleware),
-})
-
-setupListeners(store.dispatch)
